regen petstore code
