--- petstore/clients/javascript/src/models/index.ts
@@ -1,20 +1,0 @@
-// Licensed under the MIT License.
-
-export {
-  Insurance,
-  PetStoreError,
-  InsuranceUpdate,
-  CheckupUpdate,
-  Checkup,
-  CheckupCollectionWithNextLink,
-  Owner,
-  OwnerUpdate,
-  OwnerCreate,
-  OwnerCollectionWithNextLink,
-  Toy,
-  ToyCollectionWithNextLink,
-  Pet,
-  PetUpdate,
-  PetCreate,
-  PetCollectionWithNextLink,
-} from "./models.js";
