+++ petstore/clients/javascript/src/petStoreClient.ts
@@ -1,35 +1,35 @@
 // Licensed under the MIT License.
 
 import {
-  getOwnerInsuranceOperations,
+  _getOwnerInsuranceOperations,
   OwnerInsuranceOperations,
 } from "./classic/ownerInsurance/index.js";
 import {
-  getOwnerCheckupsOperations,
+  _getOwnerCheckupsOperations,
   OwnerCheckupsOperations,
 } from "./classic/ownerCheckups/index.js";
 import {
-  getOwnersOperations,
+  _getOwnersOperations,
   OwnersOperations,
 } from "./classic/owners/index.js";
 import {
-  getCheckupsOperations,
+  _getCheckupsOperations,
   CheckupsOperations,
 } from "./classic/checkups/index.js";
 import {
-  getToyInsuranceOperations,
+  _getToyInsuranceOperations,
   ToyInsuranceOperations,
 } from "./classic/toyInsurance/index.js";
-import { getToysOperations, ToysOperations } from "./classic/toys/index.js";
+import { _getToysOperations, ToysOperations } from "./classic/toys/index.js";
 import {
-  getPetInsuranceOperations,
+  _getPetInsuranceOperations,
   PetInsuranceOperations,
 } from "./classic/petInsurance/index.js";
 import {
-  getPetCheckupsOperations,
+  _getPetCheckupsOperations,
   PetCheckupsOperations,
 } from "./classic/petCheckups/index.js";
-import { getPetsOperations, PetsOperations } from "./classic/pets/index.js";
+import { _getPetsOperations, PetsOperations } from "./classic/pets/index.js";
 import {
   createPetStore,
   PetStoreContext,
@@ -57,15 +57,15 @@ export class PetStoreClient {
       userAgentOptions: { userAgentPrefix },
     });
     this.pipeline = this._client.pipeline;
-    this.ownerInsurance = getOwnerInsuranceOperations(this._client);
-    this.ownerCheckups = getOwnerCheckupsOperations(this._client);
-    this.owners = getOwnersOperations(this._client);
-    this.checkups = getCheckupsOperations(this._client);
-    this.toyInsurance = getToyInsuranceOperations(this._client);
-    this.toys = getToysOperations(this._client);
-    this.petInsurance = getPetInsuranceOperations(this._client);
-    this.petCheckups = getPetCheckupsOperations(this._client);
-    this.pets = getPetsOperations(this._client);
+    this.ownerInsurance = _getOwnerInsuranceOperations(this._client);
+    this.ownerCheckups = _getOwnerCheckupsOperations(this._client);
+    this.owners = _getOwnersOperations(this._client);
+    this.checkups = _getCheckupsOperations(this._client);
+    this.toyInsurance = _getToyInsuranceOperations(this._client);
+    this.toys = _getToysOperations(this._client);
+    this.petInsurance = _getPetInsuranceOperations(this._client);
+    this.petCheckups = _getPetCheckupsOperations(this._client);
+    this.pets = _getPetsOperations(this._client);
   }
 
   /** The operation groups for ownerInsurance */
