+++ petstore/spec/tspconfig.yaml
@@ -17,7 +17,6 @@ options:
   "@typespec/http-client-java":
     emitter-output-dir: "{project-root}/../clients/java"
   "@azure-tools/typespec-ts":
-    isModularLibrary: true
     emitter-output-dir: "{project-root}/../clients/javascript"
     packageDetails:
       name: "@unbranded/petstore"
